+++ aws-android-sdk-iot/src/main/java/com/amazonaws/services/iotdata/model/UpdateThingShadowRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -21,12 +21,12 @@ import com.amazonaws.AmazonWebServiceRequest;
 
 /**
  * <p>
- * Updates the thing shadow for the specified thing.
+ * Updates the shadow for the specified thing.
  * </p>
  * <p>
  * For more information, see <a href=
  * "http://docs.aws.amazon.com/iot/latest/developerguide/API_UpdateThingShadow.html"
- * >UpdateThingShadow</a> in the <i>AWS IoT Developer Guide</i>.
+ * >UpdateThingShadow</a> in the AWS IoT Developer Guide.
  * </p>
  */
 public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements Serializable {
@@ -37,9 +37,20 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
      * <p>
      * <b>Constraints:</b><br/>
      * <b>Length: </b>1 - 128<br/>
-     * <b>Pattern: </b>[a-zA-Z0-9_-]+<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
      */
     private String thingName;
+
+    /**
+     * <p>
+     * The name of the shadow.
+     * </p>
+     * <p>
+     * <b>Constraints:</b><br/>
+     * <b>Length: </b>1 - 64<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
+     */
+    private String shadowName;
 
     /**
      * <p>
@@ -55,7 +66,7 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
      * <p>
      * <b>Constraints:</b><br/>
      * <b>Length: </b>1 - 128<br/>
-     * <b>Pattern: </b>[a-zA-Z0-9_-]+<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
      *
      * @return <p>
      *         The name of the thing.
@@ -72,7 +83,7 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
      * <p>
      * <b>Constraints:</b><br/>
      * <b>Length: </b>1 - 128<br/>
-     * <b>Pattern: </b>[a-zA-Z0-9_-]+<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
      *
      * @param thingName <p>
      *            The name of the thing.
@@ -92,7 +103,7 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
      * <p>
      * <b>Constraints:</b><br/>
      * <b>Length: </b>1 - 128<br/>
-     * <b>Pattern: </b>[a-zA-Z0-9_-]+<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
      *
      * @param thingName <p>
      *            The name of the thing.
@@ -102,6 +113,63 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
      */
     public UpdateThingShadowRequest withThingName(String thingName) {
         this.thingName = thingName;
+        return this;
+    }
+
+    /**
+     * <p>
+     * The name of the shadow.
+     * </p>
+     * <p>
+     * <b>Constraints:</b><br/>
+     * <b>Length: </b>1 - 64<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
+     *
+     * @return <p>
+     *         The name of the shadow.
+     *         </p>
+     */
+    public String getShadowName() {
+        return shadowName;
+    }
+
+    /**
+     * <p>
+     * The name of the shadow.
+     * </p>
+     * <p>
+     * <b>Constraints:</b><br/>
+     * <b>Length: </b>1 - 64<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
+     *
+     * @param shadowName <p>
+     *            The name of the shadow.
+     *            </p>
+     */
+    public void setShadowName(String shadowName) {
+        this.shadowName = shadowName;
+    }
+
+    /**
+     * <p>
+     * The name of the shadow.
+     * </p>
+     * <p>
+     * Returns a reference to this object so that method calls can be chained
+     * together.
+     * <p>
+     * <b>Constraints:</b><br/>
+     * <b>Length: </b>1 - 64<br/>
+     * <b>Pattern: </b>[a-zA-Z0-9:_-]+<br/>
+     *
+     * @param shadowName <p>
+     *            The name of the shadow.
+     *            </p>
+     * @return A reference to this updated object so that method calls can be
+     *         chained together.
+     */
+    public UpdateThingShadowRequest withShadowName(String shadowName) {
+        this.shadowName = shadowName;
         return this;
     }
 
@@ -163,6 +231,8 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
         sb.append("{");
         if (getThingName() != null)
             sb.append("thingName: " + getThingName() + ",");
+        if (getShadowName() != null)
+            sb.append("shadowName: " + getShadowName() + ",");
         if (getPayload() != null)
             sb.append("payload: " + getPayload());
         sb.append("}");
@@ -175,6 +245,7 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
         int hashCode = 1;
 
         hashCode = prime * hashCode + ((getThingName() == null) ? 0 : getThingName().hashCode());
+        hashCode = prime * hashCode + ((getShadowName() == null) ? 0 : getShadowName().hashCode());
         hashCode = prime * hashCode + ((getPayload() == null) ? 0 : getPayload().hashCode());
         return hashCode;
     }
@@ -194,6 +265,11 @@ public class UpdateThingShadowRequest extends AmazonWebServiceRequest implements
             return false;
         if (other.getThingName() != null
                 && other.getThingName().equals(this.getThingName()) == false)
+            return false;
+        if (other.getShadowName() == null ^ this.getShadowName() == null)
+            return false;
+        if (other.getShadowName() != null
+                && other.getShadowName().equals(this.getShadowName()) == false)
             return false;
         if (other.getPayload() == null ^ this.getPayload() == null)
             return false;