+++ aws-android-sdk-iot/src/main/java/com/amazonaws/services/iotdata/model/transform/GetThingShadowRequestMarshaller.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2010-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -55,6 +55,10 @@ public class GetThingShadowRequestMarshaller implements
                 "{thingName}",
                 (getThingShadowRequest.getThingName() == null) ? "" : StringUtils
                         .fromString(getThingShadowRequest.getThingName()));
+        if (getThingShadowRequest.getShadowName() != null) {
+            request.addParameter("name",
+                    StringUtils.fromString(getThingShadowRequest.getShadowName()));
+        }
         request.setResourcePath(uriResourcePath);
         if (!request.getHeaders().containsKey("Content-Type")) {
             request.addHeader("Content-Type", "application/x-amz-json-1.0");