+++ aws-android-sdk-s3-test/src/androidTest/java/com/amazonaws/mobileconnectors/s3/transferutility/PauseTransferIntegrationTest.java
@@ -23,6 +23,7 @@ import com.amazonaws.services.s3.S3IntegrationTestBase;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.io.File;
@@ -113,7 +114,7 @@ public final class PauseTransferIntegrationTest extends S3IntegrationTestBase {
 
         // cancel early to avoid having to wait for completion
         util.cancel(observer.getId());
-        completed.await(100, TimeUnit.MILLISECONDS);
+        completed.await(1000, TimeUnit.MILLISECONDS);
     }
 
     private final class TestListener implements TransferListener {